User Images and Files
Adding user image functionality. 1.Really need to add delete functionality so they can change their user image

2. Really need to add this feature for the overall database.

3. On behalf of the database here are some things I need to look into:

A. Find a way to possibly change the form to have multiple input fields sent once and not individually.

B. Find a way to add a custom user image of the deadeye that does not display but is

I. Stored as a link on the table.
II. Stored as a seperate folder link pertaining to each post. (things might get messy hur)